+++ kytos/cli/commands/bug_report/__init__.py
@@ -1,0 +1,1 @@
+"""BugReport CLI Commands."""